Consolide github_user_aggregator
